// src/app/core/interfaces/character.ts
export interface ICharacter {
  uid: string;
  properties: {
    birth_year: string;
    created: string;
    edited: string;
    eye_color: string;
    hair_color: string;
    height: string;
    homeworld: string;
    gender: string;
    mass: string;
    name: string;
    skin_color: string;
    url: string;
  };
}